Server/requirements.txt
fastapi==0.104.1
mangum==0.17.0
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.2
openai==1.12.0
fr24sdk
